CR-2693: refactor for the kms key builder
